成功启动无报错

+++ shiyue-news/src/main/java/com/ye/shiyue/news/vo/ReturnCategory.java
@@ -3,11 +3,12 @@ package com.ye.shiyue.news.vo;
 import lombok.Data;
 
 import java.util.List;
+import java.util.Locale;
 
 @Data
 public class ReturnCategory {
 
     private List<String> categoryList;
 
-    private List<Integer> numberList;
+    private List<Long> numberList;
 }
